requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.13.0b2
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
configparser==6.0.1
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
docopt==0.6.2
Faker==12.0.1
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.42
google==3.0.0
google-ai-generativelanguage==0.5.4
google-api-core==2.17.1
google-auth==2.29.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
googletrans==3.0.0
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.4
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mixer==7.2.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
multidict==6.0.5
namex==0.0.7
newsapi==0.1.1
newsapi-cli==1.2.1
newsapi-python==0.2.7
notification==0.2.1
numpy==1.26.4
openai==0.28.0
opt-einsum==3.3.0
optree==0.10.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
playsound==1.3.0
plyer==2.1.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyjokes==0.6.0
PyMsgBox==1.0.9
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.9.3.3
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
referencing==0.34.0
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.2
speedtest==0.0.1
speedtest-cli==2.1.3
streamlit==1.33.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.1
wikipedia==1.4.0
WMI==1.5.1
wrapt==1.16.0
yarl==1.9.4
